+++ src/vw/Plate/BlobManager.cc
@@ -81,7 +81,7 @@ int vw::platefile::BlobManager::get_next_available_blob() {
   } else {
     BlobCacheRecord rec;
     m_blob_locks.push_back(rec);
-    return m_blob_locks.size() - 1;
+    return boost::numeric_cast<int>(m_blob_locks.size()) - 1;
   }
 }
 
@@ -104,7 +104,7 @@ vw::platefile::BlobManager::BlobManager(uint64 max_blob_size, int initial_nblobs
 /// Return the number of blobs currently in use.
 unsigned vw::platefile::BlobManager::num_blobs() {
   Mutex::Lock lock(m_mutex);
-  return m_blob_locks.size();
+  return boost::numeric_cast<unsigned>(m_blob_locks.size());
 }
 
 vw::int64 vw::platefile::BlobManager::max_blob_size() {
@@ -142,6 +142,6 @@ int vw::platefile::BlobManager::request_lock(uint64 &size) {
 // "committing" the write to the blob when you are finished with it.).
 void vw::platefile::BlobManager::release_lock(int blob_id, uint64 blob_offset) {
   Mutex::Lock lock(m_mutex);
-  m_blob_locks[blob_id].unlock(blob_offset);
+  m_blob_locks[blob_id].unlock(boost::numeric_cast<int>(blob_offset));
 }
